Added script to create a kickstart floppy image containing a ks file
- Updated the root README to include info about the kickstart directory
- Added a README to the kickstart directory to explain the purpose of
individual scripts
- Added gen_ks-floppy.sh which creates a floppy image that contains a
user specified kickstart file for use on iLO, DRAC, and IPMI virtual
drives. This will allow a sys admin to create portable small disks that
contain just the build instructions. As a result, the distro iso does
not need to be modified. (The creation script was tested on Mac OS X
Lion 10.7 and should work on Linux)